#2941 - Multi-value concept feature
- Extract a base class from the ConceptFeatureEditor
